优化删除已经删除的收藏记录报错

+++ src/main/java/com/wudaokou/backend/history/HistoryController.java
@@ -77,7 +77,8 @@ public class HistoryController {
     ResponseEntity<?> delete(@PathVariable String s){
         try {
             int id = Integer.parseInt(s);
-            historyRepository.deleteById(id);
+            if(historyRepository.findById(id).isPresent())
+                historyRepository.deleteById(id);
         }catch(NumberFormatException nfe){
             try{
                 Customer customer = securityRelated.getCustomer();
